+++ CHANGELOG.md
@@ -1,4 +1,7 @@
 ## 3.58.1 (Unreleased)
+### Added
+- Support for creating `oci_database_autonomous_database` resource by cloning from a backup of an existing Autonomous Database.
+
 ## 3.58.0 (January 15, 2020)
 
 ### Added